include scores in properties, refactor scores and fix circular import factories and guacamole
