+++ src/gt4sd/properties/scores/core.py
@@ -22,9 +22,8 @@
 # SOFTWARE.
 #
 """Implementation of scorers."""
-import json
 from functools import partial
-from typing import Any, Callable, Dict, List, Tuple, Type, Union
+from typing import Any, Callable, Dict, List, Type
 
 import numpy as np
 from rdkit import Chem
@@ -350,43 +349,3 @@ class QEDScorer(TargetValueScorer):
             A score for the given SMILES
         """
         return Chem.QED.qed(Chem.MolFromSmiles(smiles))
-
-
-SCORING_FUNCTIONS = {
-    "rdkit_scorer": RDKitDescriptorScorer,
-    "tanimoto_scorer": TanimotoScorer,
-    "isomer_scorer": IsomerScorer,
-    "smarts_scorer": SMARTSScorer,
-    "qed_scorer": QEDScorer,
-}
-
-
-def get_target_parameters(
-    target: Union[str, Dict[str, Any]]
-) -> Tuple[List[Type[Any]], List[float]]:
-    """Generates a tuple of scorers and weight list
-
-    Args:
-        target: scoring functions and parameters related to it
-
-    Return:
-        A tuple containing scoring functions and weight list
-    """
-    score_list = []
-    weights = []
-    target_dictionary: Dict[str, Any] = {}
-    if isinstance(target, str):
-        target_dictionary = json.loads(target)
-    elif isinstance(target, dict):
-        target_dictionary = target
-    else:
-        raise ValueError(
-            f"{target} of type {type(target)} is not supported: provide 'str' or 'Dict[str, Any]'"
-        )
-    for scoring_function_name, parameters in target_dictionary.items():
-        weight = 1.0
-        if "weight" in parameters:
-            weight = parameters.pop("weight")
-        score_list.append(SCORING_FUNCTIONS[scoring_function_name](**parameters))
-        weights.append(weight)
-    return (score_list, weights)
